Made Component for Creating a Task
Made the component for creating a task, also created a shared form to reuse with update. Modified the index.html file to have the title reflect the name of the app. Edited the header file to have the name of the app. Included a message in the msgAlert for successful creation of a task.